+++ LatticePoly/include/MCLiqLattice.hpp
@@ -23,8 +23,8 @@ public:
 	void TrialMove(double*);
 	void AcceptMove();
 	
-	double GetCouplingEnergy(const int[Ntot]) const;
-    double GetCouplingEnergyPainter(const int[Ntot], const  double[Ntot] ) const;
+	double GetCouplingEnergy(const double[Ntot]) const;
+    double GetCouplingEnergyPainter(const double[Ntot], const  double[Ntot] ) const;
